refactor: move workflows entity store/query/routes/types into @repo/shared (#1518)
