+++ app/services/workflows/brc/routes.ts
@@ -1,0 +1,3 @@
+export const API = {
+  pangenomes: "/api/pangenomes.json",
+} as const;
