--- app/services/workflows/types.ts
@@ -1,3 +1,0 @@
-import { API } from "./routes";
-
-export type EntityRoute = keyof typeof API;
