+++ app/views/OrganismView/components/Main/components/WorkflowsSection/workflowsSection.tsx
@@ -1,4 +1,3 @@
-import { getWorkflows } from "@/services/workflows/entities";
 import { Accordion } from "@/views/AnalyzeWorkflowsView/components/Main/components/Accordion/accordion";
 import { StyledSectionTitle } from "@/views/OrganismView/components/Main/main.styles";
 import { buildOrganismWorkflows } from "@/views/OrganismView/components/Main/utils";
@@ -6,6 +5,7 @@ import { useFeatureFlag } from "@databiosphere/findable-ui/lib/hooks/useFeatureF
 import { TYPOGRAPHY_PROPS } from "@databiosphere/findable-ui/lib/styles/common/mui/typography";
 import { Stack } from "@mui/material";
 import { ROUTES } from "@repo/shared/routes/constants";
+import { getWorkflows } from "@repo/shared/services/workflows/entities";
 import { JSX } from "react";
 import { EmptyState } from "../EmptyState/emptyState";
 import { Props } from "./types";
