+++ app/views/OrganismView/hooks/UseShowPangenome/hook.ts
@@ -1,5 +1,5 @@
 import type { Pangenome } from "@/apis/catalog/brc-analytics-catalog/common/pangenome";
-import { getPangenome } from "@/services/workflows/entities";
+import { getPangenome } from "@/services/workflows/brc/entities";
 import { useFeatureFlag } from "@databiosphere/findable-ui/lib/hooks/useFeatureFlag/useFeatureFlag";
 
 /**
