+++ app/views/OrganismWorkflowInputsView/organismWorkflowInputsView.tsx
@@ -8,14 +8,14 @@ import { Main } from "@/components/Entity/components/ConfigureWorkflowInputs/com
 import { SideColumn } from "@/components/Entity/components/ConfigureWorkflowInputs/components/SideColumn/sideColumn";
 import { WorkflowEntityContext } from "@/components/Entity/components/ConfigureWorkflowInputs/providers/WorkflowEntity/context";
 import { buildWorkflowEntityValue } from "@/components/Entity/components/ConfigureWorkflowInputs/providers/WorkflowEntity/utils";
-import { getWorkflow } from "@/services/workflows/entities";
-import { getEntity } from "@/services/workflows/query";
 import {
   BackPageContent,
   BackPageContentSideColumn,
   BackPageHero,
   BackPageView,
 } from "@databiosphere/findable-ui/lib/components/Layout/components/BackPage/backPageView.styles";
+import { getWorkflow } from "@repo/shared/services/workflows/entities";
+import { getEntity } from "@repo/shared/services/workflows/query";
 import { JSX, useMemo } from "react";
 import { useConfigureInputs } from "../WorkflowInputsView/hooks/UseConfigureInputs/useConfigureInputs";
 import { StyledBackPageContentMainColumn } from "../WorkflowInputsView/workflowInputsView.styles";
