+++ app/views/WorkflowView/hooks/UseWorkflowEntities/utils.ts
@@ -1,6 +1,5 @@
 import type { BRCDataCatalogOrganism } from "@/apis/catalog/brc-analytics-catalog/common/entities";
 import type { GA2OrganismEntity } from "@/apis/catalog/ga2/entities";
-import { findOrganism, getAssembly } from "@/services/workflows/entities";
 import type { Organism } from "@/views/OrganismView/types";
 import { mapOrganismEntityToOrganism } from "@/views/OrganismWorkflowInputsView/utils";
 import type { Assembly } from "@/views/WorkflowInputsView/types";
@@ -8,6 +7,10 @@ import { mapAssemblyToOrganism } from "@/views/WorkflowInputsView/utils";
 import { WORKFLOW_SCOPE } from "@repo/shared/apis/schema-types";
 import { sanitizeEntityId } from "@repo/shared/apis/utils";
 import type { Workflow } from "@repo/shared/apis/workflow";
+import {
+  findOrganism,
+  getAssembly,
+} from "@repo/shared/services/workflows/entities";
 
 /**
  * Resolves the reference-assembly genome, when one is configured.
