+++ packages/shared/services/workflows/routes.ts
@@ -1,7 +1,6 @@
 export const API = {
   assemblies: "/api/assemblies.json",
   organisms: "/api/organisms.json",
-  pangenomes: "/api/pangenomes.json",
   workflowAssemblyMappings: "/api/workflow-assembly-mappings.json",
   workflows: "/api/workflows.json",
 } as const;
