+++ packages/shared/services/workflows/store.ts
@@ -1,25 +1,25 @@
-import { EntityRoute } from "./types";
+import type { EntityStoreKey } from "./types";
 
 // entityListType -> T[]
-const ENTITIES_BY_TYPE = new Map<EntityRoute, unknown[]>();
+const ENTITIES_BY_TYPE = new Map<EntityStoreKey, unknown[]>();
 
 // entityListType -> entityId -> T
-const ENTITIES_BY_ENTITY_ID = new Map<EntityRoute, Map<string, unknown>>();
+const ENTITIES_BY_ENTITY_ID = new Map<EntityStoreKey, Map<string, unknown>>();
 
 /**
  * Gets entities by entity id.
  * @returns Map of entity list types to entity id to entity.
  */
-export function getEntitiesById(): Map<EntityRoute, Map<string, unknown>> {
-  return ENTITIES_BY_ENTITY_ID as Map<EntityRoute, Map<string, unknown>>;
+export function getEntitiesById(): Map<EntityStoreKey, Map<string, unknown>> {
+  return ENTITIES_BY_ENTITY_ID as Map<EntityStoreKey, Map<string, unknown>>;
 }
 
 /**
  * Gets entities by entity list type.
  * @returns Map of entity list types to entities.
  */
-export function getEntitiesByType(): Map<EntityRoute, unknown[]> {
-  return ENTITIES_BY_TYPE as Map<EntityRoute, unknown[]>;
+export function getEntitiesByType(): Map<EntityStoreKey, unknown[]> {
+  return ENTITIES_BY_TYPE as Map<EntityStoreKey, unknown[]>;
 }
 
 /**
@@ -28,7 +28,7 @@ export function getEntitiesByType(): Map<EntityRoute, unknown[]> {
  * @param entitiesById - Map of entity id to entity.
  */
 export function setEntitiesById(
-  entityListType: EntityRoute,
+  entityListType: EntityStoreKey,
   entitiesById: Map<string, unknown>
 ): void {
   ENTITIES_BY_ENTITY_ID.set(entityListType, entitiesById);
@@ -40,7 +40,7 @@ export function setEntitiesById(
  * @param entities - Entities.
  */
 export function setEntitiesByType(
-  entityListType: EntityRoute,
+  entityListType: EntityStoreKey,
   entities: unknown[]
 ): void {
   ENTITIES_BY_TYPE.set(entityListType, entities);
