+++ tests/services/workflows.store.test.ts
@@ -3,7 +3,7 @@ import {
   getEntitiesByType,
   setEntitiesById,
   setEntitiesByType,
-} from "@/services/workflows/store";
+} from "@repo/shared/services/workflows/store";
 
 describe("workflows store", () => {
   beforeEach(() => {
